vim-vimwiki: update to 2.0.1
This includes patching a script to convert CamelCase links.

The script is taken from
https://groups.google.com/forum/?fromgroups=#!topic/vimwiki/xChU93ovXXE
and is under the GPL3

I patched this quite a bit (patch added, rather than changing directly)
in order to fix some bugs/problems.
It should now be safe to call it multiple times and it should
link less things that are not supposed to get links.